chore: Refactor live region in progress bar
